requirements.txt
absl-py==0.12.0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.0
colorama==0.4.4
commonmark==0.9.1
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2
entrypoints==0.3
Flask==2.0.0
flatbuffers==1.12
flickrapi==2.4.0
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
image==1.5.33
imageio==2.9.0
incremental==21.3.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==2.0.0
jedi==0.18.0
Jinja2==3.0.0
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
priority==1.3.0
progress==1.5
progress-bar==8
progressbar==2.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
Protego==0.1.16
protobuf==3.17.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
queuelib==1.6.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rich==6.1.2
rsa==4.7.2
scipy==1.6.3
Scrapy==2.5.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.5.0
torch==1.8.1
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
typing-extensions==3.7.4.3
urllib3==1.25.11
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.3.0
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zope.interface==5.4.0
